requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
redis
rq
python-dotenv
marshmallow
kubernetes
gunicorn
alembic
httpx